Manager and main updated

+++ model/src/Manager.java
@@ -1,18 +1,18 @@
 public class Manager extends Employee{
 
-    private Float salary;
+    private Double salary;
 
-    public Manager(String id, String name, String surname, Float salary) {
+    public Manager(String id, String name, String surname, Double salary) {
         super(name, surname);
         this.salary =salary;
 
     }
 
-    public Float getSalary() {
+    public Double getSalary() {
         return salary;
     }
 
-    public void setSalary(Float salary) {
+    public void setSalary(Double salary) {
         this.salary = salary;
     }
